# lib/api-spec/openapi.yml
openapi: 3.1.0
info:
  title: TestCaseHub API
  version: "1.0.0"
  description: UAT Test Case Management API
servers:
  - url: http://localhost:3000/api
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TestStep:
      type: object
      properties:
        id: { type: integer }
        test_case_id: { type: integer }
        step_number: { type: string }
        instruction: { type: string }
        test_data: { type: string, nullable: true }
        expected_result: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...
    TestRun:
      type: object
      properties:
        id: { type: integer }
        project_id: { type: integer }
        name: { type: string }
        status: { type: string, enum: [scheduled, in_progress, completed] }
        scheduled_at: { type: string, format: date-time, nullable: true }
        passed: { type: boolean, nullable: true }
        source_test_run_id: { type: integer, nullable: true }
        entry_confirmed: { type: boolean }
        entry_confirmed_by_user_id: { type: integer, nullable: true }
        entry_confirmed_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
paths:
...
  /test-cases/{testCaseId}/steps:
    get:
      summary: List steps for test case
      tags: [TestSteps]
      parameters:
        - in: path
          name: testCaseId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestStep"
    post:
      summary: Create step
      tags: [TestSteps]
      parameters:
        - in: path
          name: testCaseId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestStep"
      responses:
        "201":
          description: Created
  /test-cases/{testCaseId}/steps/bulk:
    post:
      summary: Bulk create steps
      tags: [TestSteps]
      parameters:
        - in: path
          name: testCaseId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/TestStep"
      responses:
        "201":
          description: Created
  /test-steps/{stepId}:
    put:
      summary: Update step
      tags: [TestSteps]
      parameters:
        - in: path
          name: stepId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestStep"
      responses:
        "200":
          description: OK
    delete:
      summary: Delete step
      tags: [TestSteps]
      parameters:
        - in: path
          name: stepId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
...
  /test-runs/{testRunId}:
    get:
      summary: Get test run by id
      tags: [TestRuns]
      parameters:
        - in: path
          name: testRunId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
    patch:
      summary: Update test run
      tags: [TestRuns]
      parameters:
        - in: path
          name: testRunId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestRun"
      responses:
        "200":
          description: OK
    delete:
      summary: Delete test run (scheduled only)
      tags: [TestRuns]
      parameters:
        - in: path
          name: testRunId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
...
  /test-runs/{testRunId}/re-run:
    post:
      summary: Re-run failed test cases
      tags: [TestRuns]
      parameters:
        - in: path
          name: testRunId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                scheduledAt: { type: string, format: date-time }
                failedOnly: { type: boolean }
      responses:
        "201":
          description: Created
  /projects/{projectId}/test-runs/analytics:
    get:
      summary: Get test run analytics for project
      tags: [TestRuns]
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /test-runs/{testRunId}/full-report:
    get:
      summary: Get full test run report
      tags: [TestRuns]
      parameters:
        - in: path
          name: testRunId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /test-runs/{testRunId}/confirm-entry:
    patch:
      summary: Confirm entry criteria
      tags: [TestRuns]
      parameters:
        - in: path
          name: testRunId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /test-runs/{testRunId}/access-qr:
    get:
      summary: Generate QR code for test run
      tags: [TestRuns]
      parameters:
        - in: path
          name: testRunId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessUrl: { type: string }
                  qrDataUrl: { type: string }
...
  /dashboard/tester/{userId}/test-runs:
    get:
      summary: Get tester's assigned test runs
      tags: [Dashboard]
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
...
security:
  - BearerAuth: []